% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FSAmisc.R
\docType{package}
\name{FSAmisc}
\alias{FSAmisc}
\alias{FSAmisc-package}
\title{Miscellaneous functions for simple fisheries stock assessment methods.}
\description{
In progress or miscellaneous functions eventually destined for the FSA package.
}
\details{
In progress or miscellaneous functions eventually destined for the FSA package.

Questions, comments, or suggestions should be given on the \href{https://github.com/droglenc/FSAmisc/issues}{GitHub FSAmisc Issues page}.

Packages with related functionality by the same author are
\itemize{
  \item The \href{http://fishr.wordpress.com/fsa/fsa/}{FSA package} contains more evolved functions.
  \item The \href{http://fishr.wordpress.com/fsa/fsadata/}{FSAdata package} contains additional data sets.
  \item The \href{https://github.com/droglenc/FSAsim}{FSAsim package} simulation routines for various fisheries methods.
  \item The \href{https://github.com/droglenc/FSAWs}{FSAWs package} contains functions for developing and validating standard weight equations.
}
}

